Worked on the issue with sorting
Started with the Respository. Removed the Transaction transaction object that was not being used. Updated the try/catch of the populateData() method to be a try with resources. That way it would self close the stream. In the findById() method, I removed the redundant return statement. There was no need for local variable declaration.

Initially, the sorting was being done on the frontend with Thmyleaf's th:with. I removed this and the tBody. I moved the sorting by date to the Service layer using a stream. Now that the sorting is being done in the Service layer, I can pass a sorted list into the view.